Do not set sslverify to false
Removes all instances of sslverify being set to false. Because sslverify
defaults to true, we don't need to explicitly set it to true.

This commit also reformats the affected arrays to span multiple lines where
appropriate, for easier human parsing of future changes (e.g. only the affected
line gets changed if an element is added/removed, vs. the entire array
showing in the diff).

Closes #1550