Alter member search to be more inclusive (#1702)
* Alter member search to be more inclusive

This change mimics WP core's method of prepending and appending
`*` to user searches, so that when the `WP_User_Query` is prepared,
the `LIKE` clauses will be surrounded by `%`. This results in finding
e.g. a member with the email address `john123@gmail.com` when searching
for `john`; previously this member would not be found.

* Tabs, spaces, who knows

* tabs, not spaces

* Move our search query alteration to later on

This way we don't impact special search cases, including:
payment_profile:
first_name:
last_name: